nav bar / style
Updated nav bar to include active class and other links (like index.html)
added end of file - not sure what is going on but many of these files have a closing body tag right after the opening body tag.  this makes a mess of the HTML. I deleted the offending closing tag and added the close to the bottom with the closing html tag.
added jquery and javascript to make the menu dropdown work.